requirements-scripts.txt
boto3>=1.34.0
click>=8.1.0
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
pyyaml>=6.0.0
